Extract logging code to output.go
