+++ cmd/yb/main.go
@@ -5,13 +5,10 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
-	"sync"
 
 	"github.com/matishsiao/goInfo"
 	"github.com/spf13/cobra"
-	"github.com/yourbase/commons/envvar"
 	"github.com/yourbase/yb/internal/config"
 	"zombiezen.com/go/log"
 )
@@ -112,99 +109,4 @@ func displayOldDirectoryWarning(ctx context.Context) {
 	if _, err := os.Stat(dir); err == nil {
 		log.Warnf(ctx, "yb no longer uses %s to store files. You can remove this directory to save disk space.", dir)
 	}
-}
-
-type logger struct {
-	color      bool
-	showLevels bool
-
-	mu  sync.Mutex
-	buf []byte
-}
-
-var logInitOnce sync.Once
-
-func initLog(cfg config.Getter, showDebug bool) {
-	logInitOnce.Do(func() {
-		log.SetDefault(&log.LevelFilter{
-			Min: configuredLogLevel(cfg, showDebug),
-			Output: &logger{
-				color:      colorLogs(),
-				showLevels: showLogLevels(cfg),
-			},
-		})
-	})
-}
-
-func (l *logger) Log(ctx context.Context, entry log.Entry) {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-
-	l.buf = l.buf[:0]
-	if l.showLevels {
-		if l.color {
-			switch {
-			case entry.Level >= log.Error:
-				// Red text
-				l.buf = append(l.buf, "\x1b[31m"...)
-			case entry.Level >= log.Warn:
-				// Yellow text
-				l.buf = append(l.buf, "\x1b[33m"...)
-			default:
-				// Cyan text
-				l.buf = append(l.buf, "\x1b[36m"...)
-			}
-		}
-		switch {
-		case entry.Level >= log.Error:
-			l.buf = append(l.buf, "ERROR"...)
-		case entry.Level >= log.Warn:
-			l.buf = append(l.buf, "WARN"...)
-		case entry.Level >= log.Info:
-			l.buf = append(l.buf, "INFO"...)
-		default:
-			l.buf = append(l.buf, "DEBUG"...)
-		}
-		if l.color {
-			l.buf = append(l.buf, "\x1b[0m"...)
-		}
-		l.buf = append(l.buf, ' ')
-	}
-	l.buf = append(l.buf, entry.Msg...)
-	l.buf = append(l.buf, '\n')
-	os.Stderr.Write(l.buf)
-}
-
-func (l *logger) LogEnabled(entry log.Entry) bool {
-	return true
-}
-
-func configuredLogLevel(cfg config.Getter, showDebug bool) log.Level {
-	if showDebug {
-		return log.Debug
-	}
-	l := config.Get(cfg, "defaults", "log-level")
-	switch strings.ToLower(l) {
-	case "debug":
-		return log.Debug
-	case "warn", "warning":
-		return log.Warn
-	case "error":
-		return log.Error
-	}
-	return log.Info
-}
-
-func colorLogs() bool {
-	b, _ := strconv.ParseBool(envvar.Get("CLICOLOR", "1"))
-	return b
-}
-
-func showLogLevels(cfg config.Getter) bool {
-	out := config.Get(cfg, "defaults", "no-pretty-output")
-	if out != "" {
-		b, _ := strconv.ParseBool(out)
-		return !b
-	}
-	return !envvar.Bool("YB_NO_PRETTY_OUTPUT")
 }
